TIP-236: Remove Oro user provider
