--- src/Pim/Bundle/UserBundle/Provider/UserProvider.php
@@ -1,67 +1,0 @@
-<?php
-
-namespace Pim\Bundle\UserBundle\Provider;
-
-use Doctrine\Common\Util\ClassUtils;
-use Pim\Component\User\Repository\UserRepositoryInterface;
-use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-
-/**
- * Implementation of Symfony UserProviderInterface
- *
- * @author    Yohan Blain <yohan.blain@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class UserProvider implements UserProviderInterface
-{
-    /** @var UserRepositoryInterface */
-    protected $userRepository;
-
-    /**
-     * @param UserRepositoryInterface $userRepository
-     */
-    public function __construct(UserRepositoryInterface $userRepository)
-    {
-        $this->userRepository = $userRepository;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function loadUserByUsername($username)
-    {
-        $user = $this->userRepository->findOneByIdentifier($username);
-        if (!$user) {
-            throw new UsernameNotFoundException(sprintf('User with username "%s" does not exist.', $username));
-        }
-
-        return $user;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function refreshUser(UserInterface $user)
-    {
-        $userClass = ClassUtils::getClass($user);
-        if (!$this->supportsClass($userClass)) {
-            throw new UnsupportedUserException(sprintf('User object of class "%s" is not supported.', $userClass));
-        }
-
-        $reloadedUser = $this->loadUserByUsername($user->getUsername());
-
-        return $reloadedUser;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsClass($class)
-    {
-        return is_subclass_of($class, 'Symfony\Component\Security\Core\User\UserInterface');
-    }
-}
