+++ src/Pim/Component/User/Model/RoleInterface.php
@@ -2,12 +2,17 @@
 
 namespace Pim\Component\User\Model;
 
+use Symfony\Component\Security\Core\Role\RoleInterface as SymfonyRoleInterface;
+
 /**
  * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
  * @copyright 2018 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @todo This "write" model should not extends Symfony\Component\Security\Core\Role\RoleInterface.
+ * We should create a "read" model that implement that class.
  */
-interface RoleInterface
+interface RoleInterface extends SymfonyRoleInterface
 {
     /**
      * Return the role id
